Generates generative art as a grid of gradient colour fades

Generates generative art as a grid of gradient colour fades

Usage

gradients(nx = 5, ny = 5, s = 1234)

Arguments

nx

Number of columns. Default 5.

ny

Number of rows. Default 5.

s

Random seed. Default 1234.

Value

Tibble