Mosiac Sketch

This function generates a generative art ggplot object from voronoi tiles in a hand-sketched look.

Usage

mosaic_sketch(
  n = 10,
  fill_cols = c("#4B3F72", "#CBB3BF", "#FFC857", "#119DA4", "#19647E"),
  line_col = "white",
  bg_col = "white",
  line_size = 2,
  x_means = c(0, 10, 5),
  y_means = c(0, 7, 8),
  xy_var = 2,
  s = 1234
)

Arguments

n

Number of points to generate tiles from. Default 100.

fill_cols

Vector of colours to fill tiles with, Default c("#4B3F72", "#CBB3BF", "#FFC857", "#119DA4", "#19647E").

line_col

Colour of lines between tiles, Default "white".

bg_col

Background colour. Default "white".

line_size

Thickness of lines between tiles. Default 1.

x_means

Vector of any number of means for the x-coordinate. Default c(0, 10, 5).

y_means

Vector of any number of means for the y-coordinate. Default c(0, 7, 8).

xy_var

Numeric varaince of x and y points. Default 2.

s

Seed value. Default 1234.

Value

A ggplot object.