Sunbursts panel

This function generates a grid of sunbursts pieces

Usage

sunbursts_panel(
  n = 100,
  ncol = 4,
  nrow = 4,
  x_means = c(0, 10, 5),
  y_means = c(0, 7, 8),
  xy_var = 5,
  low = "#4e0550",
  high = "#facdfc",
  s = 1234
)

Arguments

n

Granularity. Default 100.

ncol

Number of column panels. Default 4.

nrow

Number of row panels. Default 4.

x_means

Vector of any number of means for the x-coordinate. Default c(0, 10, 5).

y_means

Vector of any number of means for the y-coordinate. Default c(0, 7, 8).

xy_var

Numeric varaince of x and y points. Default 5.

low

Colour of background. Default "#4e0550".

high

Colour of sunburst points. Default "#facdfc".

s

Seed value. Default 1234.

Value

A ggplot object.